Draft: split downgrade from Draft.py
